server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.3
bleak==0.22.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dbus-fast==2.39.6
FindMy==0.7.6
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
influxdb-client==1.48.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.2.0
propcache==0.3.0
psycopg2==2.9.10
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reactivex==4.0.4
setuptools==76.1.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.39
srp==1.0.22
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
